chore(az.sb): remove (now) unnecessary invalid-op check in az service-bus msg ctx
